fix purge: https://t.me/c/1387666944/169004

+++ pyrobot/plugins/admemes/purge.py
@@ -5,7 +5,10 @@ import asyncio
 
 from pyrogram import Client, Filters
 
-from pyrobot import COMMAND_HAND_LER
+from pyrobot import (
+    COMMAND_HAND_LER,
+    TG_MAX_SELECT_LEN
+)
 
 from pyrobot.helper_functions.admin_check import admin_check
 
@@ -13,6 +16,10 @@ from pyrobot.helper_functions.admin_check import admin_check
 @Client.on_message(Filters.command("purge", COMMAND_HAND_LER))
 async def purge(client, message):
     """ purge upto the replied message """
+    if message.chat.type not in (("supergroup", "channel")):
+        # https://t.me/c/1312712379/84174
+        return
+
     is_admin = await admin_check(message)
 
     if not is_admin:
@@ -21,20 +28,29 @@ async def purge(client, message):
     status_message = await message.reply_text("...", quote=True)
     await message.delete()
     message_ids = []
+    count_del_etion_s = 0
 
     if message.reply_to_message:
-        message_ids = list(range(
-            message.reply_to_message.message_id,
-            message.message_id
-        ))
-        await client.delete_messages(
-            chat_id=message.chat.id,
-            message_ids=message_ids,
-            revoke=True
-        )
+        for a_s_message_id in range(message.reply_to_message.message_id, message.message_id):
+            message_ids.append(a_s_message_id)
+            if len(message_ids) == TG_MAX_SELECT_LEN:
+                await client.delete_messages(
+                    chat_id=message.chat.id,
+                    message_ids=message_ids,
+                    revoke=True
+                )
+                count_del_etion_s += len(message_ids)
+                message_ids = []
+        if len(message_ids) > 0:
+            await client.delete_messages(
+                chat_id=message.chat.id,
+                message_ids=message_ids,
+                revoke=True
+            )
+            count_del_etion_s += len(message_ids)
 
     await status_message.edit_text(
-        f"deleted {len(message_ids)} messages"
+        f"deleted {count_del_etion_s} messages"
     )
     await asyncio.sleep(5)
     await status_message.delete()
